提交代码，支持经纬度查询

+++ useful-person.okrm-core/src/main/java/com/useful/person/core/domain/Schedule.java
@@ -86,14 +86,14 @@ public class Schedule {
 	 */
 	@Getter
 	@Setter
-	private float longitude;
+	private Double longitude;
 
 	/**
 	 * 纬度
 	 */
 	@Getter
 	@Setter
-	private float latitude;
+	private Double latitude;
 	/**
 	 * 描述
 	 */
